Extend the available vertical range in the "Idle" blend space.
We will use this extra space to move the coordinates for IdleUp
and IdleDown further from the center than IdleLeft and IdleRight.
This will cause the up/down points to be further from the diagonals
than the left/right points.

Doing this will fix a quirk in the Player sprite animation.
Currently the sprite animation for down-and-right displays the
Down direction. But tweaking the cordinates as explained above
will change this so that down-and-right will trigger the Right
direction, instead.